requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
bleach==1.5.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudpickle==1.6.0
colorama==0.4.3
cryptography==3.1.1
cycler==0.10.0
docutils==0.16
future==0.18.2
gast==0.2.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gym==0.17.3
gym-super-mario-bros==7.3.2
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==2.0.0
jeepney==0.4.3
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
nes-py==8.1.6
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
Pillow==7.2.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==1.9.6
pyglet==1.5.0
Pygments==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
readme-renderer==26.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scipy==1.5.2
SecretStorage==3.1.2
six==1.15.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.5.0
tensorflow-estimator==1.15.1
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tqdm==4.50.0
twine==3.2.0
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
